% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{form_input}
\alias{form_input}
\title{Formatting a scientific name}
\usage{
form_input(x)
}
\arguments{
\item{x}{A vector string representing the name of the family}
}
\value{
A dataframe containing the results
}
\description{
Formatting a scientific name: replacing the multiple white spaceas,
trim the leading the white space, capitalise the first letter
}
\examples{
form_input("   Anemone Amurensis   ")

}
