% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{show_detail}
\alias{show_detail}
\title{Search the WFO Database}
\usage{
show_detail(x = NULL, show_message = TRUE)
}
\arguments{
\item{x}{A vector string representing the name of the family}

\item{show_message}{Logical, indicating whether the validity checking should be performed and displayed}
}
\value{
A dataframe containing the results
}
\description{
Search the WorldFloraOnline Database
}
\details{
If the input is not an accepted name, the accepted will be provided.
}
\examples{
show_detail("Tsoongiodendron odorum") # A synonym
show_detail("Anemone amurensis") # A synonym
show_detail("Machilus") # A genus
show_detail("Machilus gamblei") # An accepted name
show_detail("Machilus wangchiana") # An accepted name
show_detail("Persea kadooriei") # A synonym
show_detail("Magnolia odora") # An accepted name
show_detail("Magnolia odo") # Spelling error
show_detail(c("Magnolia odo", "aaa", "Machilus gamblei")) # Spelling error
show_detail("Cyclobalanopsis neglecta")
show_detail(c(
  "Cyclobalanopsis neglecta", "Lithocarpus quercifolius",
  "Lithocarpus quercifoli", "Apple"
)) # Multiple names
show_detail("Quercus glauca") # Multiple matches
}
