% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{show_synonyms}
\alias{show_synonyms}
\title{Show synonyms of a scientific name}
\usage{
show_synonyms(x)
}
\arguments{
\item{x}{A vector string representing the name of the family}
}
\value{
A dataframe containing the results
}
\description{
Fetch synonyms of a scientific name from the WFO database
}
\details{
If the input name is not an accepted name, the accepted name will be provided.
Always pay attention to the Note.
}
\examples{
show_synonyms("Machilus wangchiana")
show_synonyms("Anemone amurensis")
show_synonyms("Tsoongiodendron odorum")
show_synonyms("Magnolia odora")
show_synonyms("Machilus chekiangensis")
show_synonyms("Lithocarpus longanoides")
show_synonyms("Quercus glauca")

}
